# docs/swagger.yaml
basePath: /
definitions:
...
  github_com_joaofilippe_subclub_internal_web_common.Response:
    properties:
      data: {}
      message:
        type: string
    type: object
...
  internal_web_subscription.PaginatedSubscriptionResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/internal_web_subscription.SubscriptionDTO'
        type: array
      totalCount:
        type: integer
    type: object
  internal_web_subscription.SubscriptionDTO:
    properties:
      createdAt:
        type: string
      customerId:
        type: string
      customerName:
        type: string
      daysUntilRenewal:
        type: integer
      id:
        type: string
      nextBillingDate:
        type: string
      nextShipmentDate:
        type: string
      planId:
        type: string
      planName:
        type: string
      shipmentStatus:
        type: string
      startDate:
        type: string
      status:
        type: string
    type: object
  internal_web_subscription.SubscriptionInputDTO:
    properties:
      customerId:
        type: string
      nextBillingDate:
        type: string
      nextShipmentDate:
        type: string
      planId:
        type: string
      shipmentStatus:
        type: string
      startDate:
        type: string
      status:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Backend application for SubClub
  title: SubClub API
  version: "1.0"
paths:
...
  /api/v1/subscriptions:
    get:
      description: Get a paginated list of subscriptions
      parameters:
      - description: Filter by Customer ID
        in: query
        name: customerId
        type: string
      - description: Filter by Status
        in: query
        name: status
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_web_subscription.PaginatedSubscriptionResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_joaofilippe_subclub_internal_web_common.Response'
      summary: List subscriptions
      tags:
      - subscriptions
    post:
      consumes:
      - application/json
      description: Creates a new subscription for a client and plan
      parameters:
      - description: Subscription data
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/internal_web_subscription.SubscriptionInputDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_web_subscription.SubscriptionDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_joaofilippe_subclub_internal_web_common.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_joaofilippe_subclub_internal_web_common.Response'
      summary: Create a subscription
      tags:
      - subscriptions
  /api/v1/subscriptions/{id}:
    delete:
      description: Soft deletes a subscription by ID
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_joaofilippe_subclub_internal_web_common.Response'
      summary: Delete a subscription
      tags:
      - subscriptions
    get:
      description: Retrieves a subscription by its UUID
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_web_subscription.SubscriptionDTO'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_joaofilippe_subclub_internal_web_common.Response'
      summary: Get a subscription by ID
      tags:
      - subscriptions
    put:
      consumes:
      - application/json
      description: Updates an existing subscription
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated subscription data
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/internal_web_subscription.SubscriptionInputDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_web_subscription.SubscriptionDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_joaofilippe_subclub_internal_web_common.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_joaofilippe_subclub_internal_web_common.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_joaofilippe_subclub_internal_web_common.Response'
      summary: Update a subscription
      tags:
      - subscriptions
...
swagger: "2.0"